Remove "Connection" singleton
With the singleton you can't use the package without already doing
stuff, e.g. loading and parsing the "database.yml". And if there is
something in "database.yml" POP doesnt like it will die on you with a
panic. So let's the user of POP decide when to actually DO stuff.